--- src/components/SearchWithConstate/HighlightedText.module.css
@@ -1,3 +1,0 @@
-.highlight {
-  background: #fcf8be;
-}
