--- src/components/SearchWithConstate/index.ts
@@ -1,2 +1,0 @@
-export { SearchInputWithSearch, SearchInput } from "./SearchInput";
-export { BulletPointListWithSearch, BulletPointList } from "./BulletPointList";
