+++ src/features/PostsWithSearchWithConstate/components/index.ts
@@ -1,0 +1,2 @@
+export { BulletPointListWithSearch } from "./BulletListWithSearch";
+export { SearchInputWithSearch } from "./SearchInput";
