+++ src/features/PostsWithSearchWithConstate/hooks/hookCreator.ts
@@ -1,0 +1,10 @@
+import constate from "constate";
+
+import useSearch, { Props, ReturnType } from "../../../hooks/useSearch";
+import { Post } from "../types";
+
+function createContexts() {
+  return constate<Props<Post>, ReturnType<Post>, []>(useSearch);
+}
+const SearchContext = createContexts();
+export { SearchContext };
