+++ src/features/PostsWithSearchWithConstate/index.ts
@@ -1,0 +1,3 @@
+export { SearchInputWithSearch, BulletPointListWithSearch } from "./components";
+export type { Post } from "./types";
+export { SearchContext } from "./hooks/hookCreator";
