+++ src/features/PostsWithSearchWithConstate/types/index.ts
@@ -1,0 +1,4 @@
+export type Post = {
+  id: number;
+  name: string;
+};
