+++ src/hooks/useSearchConstate/index.ts
@@ -1,0 +1,1 @@
+export { SearchContext } from "./constate";
